+++ Compatibility/DTrack2/DTrack2_example_with_remote_control.cpp
@@ -1,9 +1,8 @@
-/* DTrackSDK: C++ example, A.R.T. GmbH
+/* DTrackSDK in C++: DTrack2_example_with_remote_control.cpp
  *
- * example_without_remote_control:
- *    C++ example using DTrack2 with DTrack2 remote control
+ * C++ example using 'DTrack2' with DTrack2 remote control.
  *
- * Copyright (c) 2005-2017, Advanced Realtime Tracking GmbH
+ * Copyright (c) 2005-2023 Advanced Realtime Tracking GmbH & Co. KG
  * 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -28,12 +27,11 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  * 
- * Version v2.5.0
- *
  * Purpose:
  *  - example with DTrack2 remote commands: collects DTrack data
- *
+ *  - for DTrackSDK v2.5.0 (or newer)
  */
+
 #include "DTrack2.hpp"
 
 #include <string>
@@ -264,10 +262,10 @@ int main(int argc, char** argv)
 		cout << "invalid port '" << argv[2] << "'" << endl;
 		return -2;
 	}
-	
+
 	// initialize library:
-	dt = new DTrack2((const char *)argv[1], 50105, port);
-	
+	dt = new DTrack2( ( const char* )argv[ 1 ], 50105, ( unsigned short )port );
+
 	if(!dt->valid()){
 		cout << "DTrack2 init error" << endl;
 		return -3;
